# argus-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus AI — Event Collector API
  version: 2.1.0
  description: |
    REST API for the Argus AI proctoring platform.

    ## Authentication

    All endpoints (except `POST /api/v1/auth/login`) require a Bearer JWT in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via the login endpoint and expire after 24 hours.

    ## Rate Limits

    | Scope       | Limit             |
    |-------------|-------------------|
    | Global      | 5 000 req/s       |
    | Per-IP      | 100 req/s         |
    | Login       | 5 req/s (Nginx)   |

    Rate-limit responses return HTTP 429 with a `Retry-After` header.

  contact:
    name: Argus AI Engineering
    url: https://argus.ai
servers:
  - url: https://api.argus.ai
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/monitoring/sessions/{sessionId}/terminate:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Monitoring]
      summary: Terminate session
      description: |
        Immediately terminates the proctoring session. The student's exam attempt
        is flagged and cannot be continued.
      operationId: terminateSession
      responses:
        '200':
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Archive ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ─────────────────────────────────────────────────────────────
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Proctoring session identifier
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
    Forbidden:
      description: Authenticated but insufficient role/permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ActionResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
...
security:
  - BearerAuth: []